chore: extract eisuke section
